try_flask_streamlit/resume-analyzer/requirements.txt
# flask
# streamlit
# pdfplumber
# python-docx
# sentence-transformers
# scikit-learn
# pandas
# matplotlib
# openai
# reportlab
# requests

# Core
streamlit==1.26.0
flask==2.3.5

# Data
pandas==2.1.1
matplotlib==3.8.1
python-dotenv==1.0.0
requests==2.32.5
pdfplumber==0.9.0
python-docx==0.8.12

# AI / NLP
numpy<2.0           # avoid NumPy 2.x conflicts
torch==2.2.2         # compatible with sentence-transformers
transformers==4.44.0
sentence-transformers==2.2.2
huggingface-hub>=0.16.4,<1.0

# OpenAI
openai>=1.0.0,<1.5.0  # use new API

# Optional
watchdog  # for faster Streamlit reloading
reportlab==4.0.2  # for PDF report generation



